fix(tasks): keep restart formatter side-effect free
