+++ src/cli/gateway-cli/run-loop.ts
@@ -1,6 +1,5 @@
 import { randomUUID } from "node:crypto";
 import net from "node:net";
-// In-process gateway run loop, restart signaling, drain, and update respawn handling.
 import { clearRuntimeConfigSnapshot } from "../../config/runtime-snapshot.js";
 import {
   captureGatewayRestartTraceHandoff,
@@ -16,6 +15,8 @@ import { acquireGatewayLock } from "../../infra/gateway-lock.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
 import type { RuntimeEnv } from "../../runtime.js";
 import { createLazyImportLoader } from "../../shared/lazy-promise.js";
+// In-process gateway run loop, restart signaling, drain, and update respawn handling.
+import { formatActiveTaskRestartBlocker } from "../../tasks/task-restart-blocker.js";
 const gatewayLog = createSubsystemLogger("gateway");
 const LAUNCHD_SUPERVISED_RESTART_EXIT_DELAY_MS = 1500;
 const DEFAULT_RESTART_DRAIN_TIMEOUT_MS = 300_000;
@@ -509,7 +510,6 @@ export async function runGatewayLoop(params: {
             async () => {
               const {
                 abortEmbeddedAgentRun,
-                formatActiveTaskRestartBlocker,
                 getRuntimeConfig,
                 getInspectableActiveTaskRestartBlockers,
                 getActiveEmbeddedRunCount,
